+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## v2.5.0
+
+#### Enhancements:
+
+- Add support to the core library API client to optionally raise a custom exception
+  class when a non-200 (and not 204) response is returned from an API request.
+- Raise `AnalyzerAPIError` when a non-200 response is returned from the API.
+- Add SSL hash field to list of SSL fields in dictionary output for more convenient
+  integrations.
+- Add firstseen and lastseen dates to SSL Certificate records.
+- Optional support for the Pandas data analysis library. Adds as_df property to all
+  Analyzer objects to render the object as a Pandas dataframe. 
+- Add option to specify module-level date ranges with `datetime` objects for
+  easier integration with other libraries.
+
+
+
+#### Bug Fixes
+
+- `is_ip()` regex fix to avoid matching on hostnames with embedded IPs.
+- Fixed broken `available` property on summary objects.
+- Fixed missing publish date on Articles
+
+
+
+
 ## v2.4.2
 
 #### Enhancements: